requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-cors==3.0.10
gunicorn==20.1.0
gevent>=24.2.1
gevent-websocket==0.10.1
zope.event==4.6
zope.interface==6.1
azure-cognitiveservices-speech==1.43.0
deepl==1.11.0
python-dotenv==0.19.2
werkzeug==2.0.1
pydub==0.25.1
requests==2.31.0

